Move query output implementations to query-runner module
