+++ java/query/query-lambda/src/test/java/sleeper/query/lambda/SqsQueryProcessorLambdaIT.java
@@ -73,9 +73,9 @@ import sleeper.query.model.Query;
 import sleeper.query.model.QueryProcessingConfig;
 import sleeper.query.model.QuerySerDe;
 import sleeper.query.output.ResultsOutputConstants;
-import sleeper.query.output.S3ResultsOutput;
-import sleeper.query.output.SQSResultsOutput;
-import sleeper.query.output.WebSocketResultsOutput;
+import sleeper.query.runner.output.S3ResultsOutput;
+import sleeper.query.runner.output.SQSResultsOutput;
+import sleeper.query.runner.output.WebSocketResultsOutput;
 import sleeper.query.runner.tracker.DynamoDBQueryTracker;
 import sleeper.query.runner.tracker.DynamoDBQueryTrackerCreator;
 import sleeper.query.runner.tracker.WebSocketQueryStatusReportDestination;
