+++ java/query/query-runner/src/main/java/sleeper/query/runner/tracker/WebSocketQueryStatusReportDestination.java
@@ -22,7 +22,7 @@ import org.slf4j.LoggerFactory;
 import sleeper.query.model.LeafPartitionQuery;
 import sleeper.query.model.Query;
 import sleeper.query.output.ResultsOutputInfo;
-import sleeper.query.output.WebSocketOutput;
+import sleeper.query.runner.output.WebSocketOutput;
 import sleeper.query.tracker.QueryStatusReportListener;
 
 import java.io.IOException;
